requirements.txt
asgiref==3.9.1
Django==5.1.4
tzdata==2025.2
whitenoise==6.6.0
pymongo==4.6.0
django-environ==0.12.0
gunicorn==21.2.0